Warning fixes and refactoring
Fix various warnings.

In some places, statements of the form \x -> case x of { a -> b; ... }
have been replaced by \case { a -> b; ... }, which is
syntactic sugar enabled by the LambdaCase extension.
LambdaCase has been added to the default-extensions
field in the cabal file.

In some places, type variables in annotations in the function body
obviously intended to refer to type variables in the top-level
annotation, in the spirit of ScopedTypeVariables, but without the
explicit forall required by ScopedTypeVariables to bring top-level
type variables into scope in the function body. Consider this example:

fieldAllValues :: Table t => (forall a. Columnar' f a -> b) -> t f -> [b]
fieldAllValues  (f :: forall a. Columnar' f a -> b) (tbl :: table f) = ...

Here, the type variables b and f in the type annotation for f and the
type variables table and f in the type annotation for tbl clearly refer
to the type variables of the same name in the top-level annotation
for the function itself. Adding a forall t f b. before Table t ...
is required to bring the latter variables into scope for the function body.

(Unrelated to the missing universal qualifier, in this case, the naming
of the "table" type variables was a bit inconsistent. tbl :: table f
should instead refer to t, i.e. tbl :: t f. This naming inconsitency
was also resolved in the relevant places, however the naming of
table/t across functions/modules remains inconsistent.)