PeekField: Enable propagating backend errors
Instead of PeekField containing a function that is passed
Nothing in case of an error, change the function's type to
"Either [SomeException] a -> f" to allow passing a description
of the error(s) to the backend library.

This allows backends to throw a more descriptive exception
e.g. in case of errors during field parsing.